Fixed assembly loading issues.
