+++ Remora.Plugins/Services/PluginServiceOptions.cs
@@ -22,7 +22,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.IO;
 using Remora.Results;
 
 namespace Remora.Plugins.Services;
@@ -34,7 +33,7 @@ namespace Remora.Plugins.Services;
 /// <param name="ScanAssemblyDirectory">
 /// Whether the directory of the entry assembly should be scanned for plugins.
 /// </param>
-/// <param name="Filter">The filter to use with the plugin service's <see cref="FileSystemWatcher"/>.</param>
+/// <param name="Filter">The filter to use to find plugins.</param>
 /// <param name="ErrorDelegate">The delegate used to process plugin initialization or migration errors.</param>
 public record PluginServiceOptions
 (
